Finalizado flightstats
-Finalizada API de flightstats (Salidas/Llegadas)
-Actualizados los scripts en la instancia EC2
-Introducida medida preventiva por si la API devuelve una lista vacia

+++ Python_Scripts/Aeropuerto/mongo_insert_llegadas_salidas.py
@@ -1,8 +1,7 @@
 #Importamos las librerias de mongo, y las listas auxiliares de los Scrapers de llegadas y salidas
 import pymongo
 import urllib.parse
-from .scraper_llegadas_aeropuerto import vuelosFormateados as llegadas
-from .scraper_salidas_aeropuerto import vuelosFormateados as salidas
+from .flightstats_API import getArrivals, getDepartures
 
 #Funcion para la conexion a MongoDB, seleccion de las colecciones correspondientes, borrado de datos antiguos e insercion de datos nuevos
 def insert():
@@ -10,8 +9,13 @@ def insert():
 	dbSalidas = client.TPA.salidasAeropuerto
 	dbLlegadas = client.TPA.llegadasAeropuerto
 
-	dbSalidas.delete_many({})
-	dbLlegadas.delete_many({})
+	salidas = getDepartures()
+	llegadas = getArrivals()
 
-	dbSalidas.insert_many(salidas)
-	dbLlegadas.insert_many(llegadas)
+	if salidas:
+		dbSalidas.delete_many({})
+		dbSalidas.insert_many(salidas)
+
+	if llegadas:
+		dbLlegadas.delete_many({})
+		dbLlegadas.insert_many(llegadas)
